backend/requirements.txt
flask
mysql-connector-python
python-dotenv
flask-cors
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
flask-jwt-extended
bcrypt
flask-limiter
